[FEAT] Manitto View 내부 Configuration 구현(#430)
